requirements.txt
flask==3.1.*
flask-cors
python-dotenv
websockets
aiohttp
requests
copernicusmarine
xarray
netCDF4
numpy
apscheduler
scipy
